Refactor(auth):change Auth layout & route -Add forget password btn -Change auth layout route
