--- src/app/auth/login/page.tsx
@@ -1,7 +1,0 @@
-import LoginForm from "@/components/layouts/auth/LoginForm";
-
-function page() {
-  return <LoginForm />
-}
-
-export default page;
