+++ src/components/layouts/auth/SignUpForm.tsx
@@ -1,11 +1,11 @@
 "use client"
-import AuthFormLayout, { inputStyles } from '@/app/auth/layout'
 import MainBtn from '@/components/atoms/buttons&links/MainBtn'
 import { TextField } from '@/components/atoms/inputFields/TextFields'
 import useDisclosure from '@/hooks/useDisclosure'
 import { AuthFormProps_T } from '@/types/auth.t'
 import EyeFilledIcon from '../../../../public/icons/svgs/EyeFilledIcon'
 import EyeSlashFilledIcon from '../../../../public/icons/svgs/EyeSlashFilledIcon'
+import AuthFormLayout, { inputStyles } from './AuthFormLayout'
 function SignUpForm({toggleSignUp}:AuthFormProps_T) {
   const {isOpen:isVisible,toggle} = useDisclosure()
   return (
